Allow for moving between rounds freely.
The trick was that PendingBids were storing their own state, and thus not updating when, say, the round number changed.
Also, there is a major bug which essentially compromises data integrity if a player is added and then the app moves to rounds before the player existed. A fix just occurred to me as I was writing this commit message.